[cross-project-tests][DebugInfo] Make simplified-template-names test runnable on Darwin
